Fix minor issues

+++ repos/system_upgrade/el7toel8/actors/selinux/selinuxcontentscanner/actor.py
@@ -17,7 +17,7 @@ SED_COMMAND='/' + '\|'.join(REMOVED_TYPES_) + '/s/^/;/g'
 CONTAINER_TYPES="|".join(["container_connect_any","container_runtime_t","container_runtime_exec_t","spc_t","container_auth_t","container_auth_exec_t","spc_var_run_t","container_var_lib_t","container_home_t","container_config_t","container_lock_t","container_log_t","container_runtime_tmp_t","container_runtime_tmpfs_t","container_var_run_t","container_plugin_var_run_t","container_unit_file_t","container_devpts_t","container_share_t","container_port_t","container_build_t","container_logreader_t","docker_log_t","docker_tmpfs_t","docker_share_t","docker_t","docker_lock_t","docker_home_t","docker_exec_t","docker_unit_file_t","docker_devpts_t","docker_config_t","docker_tmp_t","docker_auth_exec_t","docker_plugin_var_run_t","docker_port_t","docker_auth_t","docker_var_run_t","docker_var_lib_t","container_domain","container_net_domain"])
 
 
-def checkModule(name)
+def checkModule(name):
     """ Check if module contains one of removed types.
         If so, comment out corresponding lines and return them.
     """
@@ -51,9 +51,12 @@ def getSelinuxModules():
 
     # modules need to be extracted into cil files
     # cd to /tmp/selinux and save working directory so that we can return there
-    wd = os.getcwd()
-    os.mkdir("/tmp/selinux")
-    os.chdir("/tmp/selinux")
+    try:
+        wd = os.getcwd()
+        os.mkdir("/tmp/selinux")
+        os.chdir("/tmp/selinux")
+    except OSError:
+        pass
 
     for (name, priority) in modules:
         if priority == "200":
@@ -91,9 +94,16 @@ def getSelinuxModules():
 
     # clean-up
     for (name, priority) in modules:
-        os.remove(name + ".cil")
-    os.rmdir("/tmp/selinux")
-    os.chdir(wd)
+        if priority not in ["100","200"]:
+            try:
+                os.remove(name + ".cil")
+            except OSError:
+                continue
+    try:
+        os.rmdir("/tmp/selinux")
+        os.chdir(wd)
+    except OSError:
+        pass
 
     return semodule_list
